backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache and session
redis==4.6.0
aioredis==2.0.1

# Background tasks
celery[redis]==5.3.4

# Security and auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.38.0

# Additional dependencies
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
requests==2.31.0
